Add workflow checking for updates to Gatsby beyond 5.11.0
Add a new `dependency-check` workflow to check whether the installed
`gatsby` version has been changed beyond v5.11.0.  If so, display a
warning about it violating Engineering Advisory 2024-03-14, and fail
the workflow check.

If the dependency is a valid version and a prior warning comment is
detected, update that comment to note that the issue has been resolved.